refactor(core): control flow migration cleanup
This adds a console warning to make it clear to users the migration is developer preview. It also shuffles around some code for better organization.
